refactor(concept:changeRequest): move header and breadcrumbs
