+++ libs/utils/src/lib/language/nb.ts
@@ -259,6 +259,7 @@ export const nb = {
     proposedChange: 'Forslag til ny {fieldType} på {lang}',
     noValue: 'Ingen verdi',
     noChangeRequestsFound: 'Her var det ingen endringsforslag',
+    edit: 'Rediger',
   },
 
   comment: {
